separate psiflow reference handling code from actual reference code
