+++ psiflow/reference/__init__.py
@@ -1,4 +1,4 @@
-from ._cp2k import CP2K  # noqa: F401
-from ._dftd3 import D3  # noqa: F401
-from .gpaw_ import GPAW  # noqa: F401
+from .cp2k import CP2K  # noqa: F401
+from .dftd3 import D3  # noqa: F401
+from .gpaw import GPAW  # noqa: F401
 from .reference import Reference, evaluate  # noqa: F401
